Refactor useFetch
